Move patching outside git repository
Downloading and applying patches in a git controlled (sub)directory
will not work with files not under git control.

This was attemped fixed with setting --git-dir to
an SDK directory, but the SDK directory has a directory called config
which git interprets as its config file.

This commit moves the downloading and applying of patches to SDK
to a directory not controlled by git. A check is added to make sure
the directory is git controlled directory.